MB-40669 Couchbase server shouldn't require home directory
In a recent change the erlang cookie was no longer passed as an argument to
ns_couchdb as it would be visible via commands such as 'ps'.  Instead
the cookie is written to a file and the name of the file was passed as
an argument.

Because ns_couchdb was no longer being started with a -setcookie argument the
erlang runtime creates a cookie and writes it to the .erlang_cookie file
in the user's home directory.  This effectively introduced a requirement
for a writable home directory which is not always the case (e.g. couchbase
is installed on a system where the couchbase user already exists but doesn't
 have a home dir).

The fix is to start ns_couchdb with "-setcookie nocookie".  This will
prevent the erlang runtime from creating a cookie and writing to the
file.  During ns_couchdb start it will read the cookie file whose name
was passed as an argument and do a erlang:set_cookie() to overwrite the
prior "nocookie" value.

Change-Id: Iad8d33acca74ed2c4917240b3c09bf1a60d4519b
Reviewed-on: http://review.couchbase.org/c/ns_server/+/133397
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>