MB-41554: Don't reuse TLS sessions in LDAP on-demand validation
As a bonus this change provides ability to pass arbitrary TLS options
when connecting to LDAP server. This can be useful when we need to:
1) comply with some security policy (set tls vsn, ciphers, etc);
2) work around some ssl bugs;
3) help clients to debug tls related issues.

The extra_tls_opts option is not available for change via API
for now, as format for it is not clear. The intented usage is
through diag/eval. Later we might want to officialy support it.
We will need to implement the parse function for it then.

Example of settings arbitrary tls opts:

curl -X POST -u user:pass 127.0.0.1:8091/diag/eval -d \
"ns_config:set_sub(ldap_settings, extra_tls_opts, "\
"[{versions, ['tlsv1.2']}])."

Change-Id: I2eade02d764b1f036a2846cd6aa97b21e0d5ff40
Reviewed-on: http://review.couchbase.org/c/ns_server/+/136559
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>