MB-48047:[BP] Add endpoint to disable unused cb_dist listeners
In some situations, when changing net settings we don't know
the previous state of the system. It makes disabling of unused
interfaces problematic. The new api just closes all the listeners
that are in use. It will be used by UI for correct node
initialization and by CLI for changing ip-family and n2n encryption.

Backports change related to MB-42373.

Reviewed-on: http://review.couchbase.org/c/ns_server/+/142449

Change-Id: Ic6a66d72df4a15a923497035eab462191a83b82a
Reviewed-on: http://review.couchbase.org/c/ns_server/+/164944
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Restriction Checker