MB-50289: Refactor cb_epmd:port_please code in order to ...
... make it possible to get the port for an arbitrary address family
and encryption setting. Introduce cb_epmd:get_port/4 function.

port_please is used by dist modules for geting the port of a remote
node for currently cofigured settings (afamily and encryption).
But for self testing purposes we want to be able to extract TLS port
of a node even when encryption is disabled, and vise versa,
to extract TCP port when encryption is enabled.

Reviewed-on: http://review.couchbase.org/c/ns_server/+/146779
Change-Id: I540f8fcda4ec107011b8054e11e1447389c1ecbe
Reviewed-on: https://review.couchbase.org/c/ns_server/+/168559
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker