MB-50289: Ignore wrong connection type when ...
... checking reestablished connection after afamily or encryption
settings changes.

When connection between nodes is dropped after reconfiguration,
both nodes start reconnecting to each other. Since the remote node
might not have changed its configuration yet, it might be still using
the "old" settings (old address family or old encryption), while we
want to verify that connection can be established using the "new"
settings. Previous logic was the following: if we see that
established connection is of a wrong type we try reconnecting, with
max retries number = 10. In practice, this algorithm doesn't always
work as TCP connections has more chances to be established faster
than TLS connections. Hence the disabling of the check.
Connectivity should be verify separately when external port is being
opened - it's done in a separate commit.

Backport of MB-44372.
Reviewed-on: http://review.couchbase.org/c/ns_server/+/146782

Change-Id: I509e122a00123348720b78d8c08befeb020987ff
Reviewed-on: https://review.couchbase.org/c/ns_server/+/168562
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker