MB-44022 Obfuscate sessionid in login expired event
Prior to this fix the sessionid in the login expired audit event was not
getting obfuscated. The reason is that session_expired doesn't have an
associated request and so information is specified in Params. When the
audit event is handled by prepare/2 the "body" is created with the
sessionid and then appended with the Params. As a result the sessionid
would be in the property list twice.

As a fix the caller supplied Params take precedence and thus there will
no longer be multiple sessionids.

Change-Id: I4dcdcd2cf9a87583d55fedf7bcacfbcab1e7c60f
Reviewed-on: http://review.couchbase.org/c/ns_server/+/148089
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>