MB-45853 Fix flushing of message
The use of misc:flush with a pattern such as

    misc:flush({retry_rename_and_refresh, _, _})

does not match messages where the first param is "message_name" and the
second and third are wild-cards. A pattern cannot be passed to a
function as they are not first class...only a value can be passed. But
the use of cut implicitly creates a value (incorrect as it almost surely
won't match the real params in the message). It ends up being

    misc:flush(fun (A, B) -> {retry_rename_and_refresh, A, B} end).

Instead pass the pattern to the ?flush macro where patterns are valid.

Change-Id: If53e50362e45e11c4c646e3dee568921671361f6
Reviewed-on: http://review.couchbase.org/c/ns_server/+/153510
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>