MB-46633 Revert part of "Patch for toy-build"
"Patch for toy-build" removed a logline (1) which printed the pre-failover
vbmap and post-failover vbmap. It was intitally deemed not necessary
since the similar data would be printed (2) when the vbmap was added via the
config backend. But that other log (2) can be truncated and to make sure we
don't loose debuggablitiy in this area this patch will re-add logline (1).

Change-Id: I1ceb442ec3a43aaf2efff099a0fa40ca41b84f93
Reviewed-on: http://review.couchbase.org/c/ns_server/+/155125
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Tested-by: <hareen.kancharla@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>