MB-39514: Fix disk usage analyzer and add alert
Fixes problems associated with the disk usage (df) analysis program
hanging and causing other problems throughout the system. This is
addressed by creating an ETS table, where results will be published,
ordered by the monotonic time as the key that can serve results even
if the main process is blocked waiting for df. Then we use the time as
a way to estimate "staleness" s/t we can have alerts to notify the
user if there is a problem that needs to be addressed. To address the
problem, all you need to do is kill the hung df if possible and
ns_disksup will attempt it again. If the process cannot be killed, the
only way to resolve the issue is likely a full restart of the OS.
Killing the processes or similar actions were all deemed too
complicated and outside of the scope of this change. The main reason
for the change, is to maximize downsides while not engaging in risky
operations such as killing spawning multiple workers.

Change-Id: I7b31ad37309b3fe58ee1766ed510376476bb9663
Reviewed-on: http://review.couchbase.org/c/ns_server/+/155868
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>