[cbcollect_info] Don't open and close files needlessly.
This was added because we didn't want to keep open files to things
like chronicle logs and snapshots, which can lead to issues on
Windows. But we never write to those files. With this patch, those
files we write to will have a cached file object, while files that we
collect via collect_file/collect_dir/CollectFile/CollectDir will not
have an open file object.

Change-Id: Ic53ee991cb92426dcd7cbe544222187288947489
Reviewed-on: http://review.couchbase.org/c/ns_server/+/157902
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>