MB-35881 Limit DCP connection names to 200 chars max
The DCP connection name consists of two node names, the name of the
bucket, and a bit of overhead. If the generated name is longer than 200
chars then we use an abbreviated name.

The abbreviated name omits the "irrelevant" node from the connection
name. On the consumer, we know which node we are at, so we don't need the
consumer node; on the producer – we don't need the producer node. A hash
of the un-abbreviated connection name is added to guarantee uniqueness.

As an example the connection names used might be:

        replication:producer_node_name->:bucket_name:<hash>
and
        replication:->consumer_node_name:bucket_name:<hash>

And to repeat, this abbreviated name is only used if the name generated
using the pre-Neo method is longer than 200 characters.

Change-Id: I59cf0f58cfb539022d2b9ffa682cd635ff6a98f9
Reviewed-on: http://review.couchbase.org/c/ns_server/+/158400
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>