MB-43151: Add option for cluster encryption level in Security page.
Cluster encryption level is visible in Security page and Servers page.
It can be modified in Security page.

This change requires regrouping of tabs from Security page.
Root Certificate and Client Certificate are grouped into one tab - Certificates.
Log Redaction, Session and the new option are grouped into one tab - Other Settings.

Cluster encryption level is visible only in a cluster that was
created with the 'node 2 node encryption' option activated.
For a cluster where 'node-to-node encryption' is not activated,
cluster encryption option looks like a disabled select input
with 'none' option selected.

For a cluster with 'node-to-node encryption' activated,
the cluster encryption level can also be seen on the Servers page,
on the details of every server, like this:
'Node-to-Node Encryption: enabled(control)'

The required permissions for a user to see this option are:
 - the cluster should be enterprise
 - cluster.admin.security.read should be true
   (whole Security page has this requirement to be visible)
The required permissions for a user to modify this option are:
 - cluster.admin.security.write

Change-Id: I67713fd9c4d01bf74b68a0949376ab977dc403a2
Reviewed-on: http://review.couchbase.org/c/ns_server/+/158439
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>