MB-35881 Limit DCP connection names to 200 chars max
The DCP connection name consists of two node names, the name of the
bucket, and a bit of overhead. If the generated name is longer than 200
chars then we use an abbreviated name.

The abbreviated name removes the common prefix between the two node
names while keeping a minimal length.

As the abbreviated connection name may no longer be unique a hash is
appended to the end.

Change-Id: I1675c72d1bc8dac1c7c39202f45c971777ce8902
Reviewed-on: http://review.couchbase.org/c/ns_server/+/159432
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>