MB-48084: Always annouce cluster_encryption_level on start.
When a node is removed from cluster, the cluster_encryption_level key is
deleted from the ns_config. Before this patch, we would only generate a change
event when the key is present in the config. We would miss any changes
to cluster_encryption_level in this case. Therefore, always generate a
change message for cluster_encryption_level on startup.

Change-Id: Iad43c8b8f621bdd383fcb52251a4195850d9e075
Reviewed-on: http://review.couchbase.org/c/ns_server/+/159921
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>