MB-51208: get_hostnames should not assume the whole cluster is EE
The get_hostnames function checks if the cluster is EE, and if so
it builds nodenames using regular rest port and ssl port. But this
is incorrect, because the cluster can contain both EE and CE nodes
simultaneously. Any attempt to build a hostname with ssl port for a
CE node when executed in a EE node will crash.

Change-Id: I34603eb9d9789433529cbf57f216ef2b30a5cbb1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/171613
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>