MB-49705 [BP] Don't leak sensitive information (engage_cluster)
Prior to this change a crash while doing an engage_cluster could leak
sensitive information such as PRIVATE KEYs. We know wrap the list of
arguments used by engage_cluster in a function and pass the function
instead.

This is a backport of MB-48222.

Change-Id: I5b268d525f6c58114fbe169d9ba7fc3f75845a9b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/172372
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>