MB-51752 [BP] Gather projector profiles when TLS is enforced
This backports these changes from MB-51197

1) Add tls status of node to dump-guts/cbcollect

Adds tls support to dump-guts, and cbcollect so that it can be used to
alter the behavior of the projector collector task and make the
correct requests to https instead of http when cluster encryption is
enabled. Also makes this information generally available for other
tasks to use in cbcollect, as needed.

2) add curl '-k' (allow-insecure) option by default to get

Allow insecure connections that don't require the proper certificate
for curl to connect to self-signed or unknown certificates on
cbcollect. Used primarily for collecting projector dumps but is also
added by default for get requests and can be expanded to other
requests in the future, if needed.

In addition the dump-guts portion was fixed to check the
cluster_encryption_level.

Change-Id: Ibbb6bc9298c74a0aea42a1c127c33e8b59203987
Reviewed-on: https://review.couchbase.org/c/ns_server/+/173500
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>