MB-52640, MB-52639: Fix broken Delta Recovery
When a bucket is brought up for the first time, ns_janitor generates
vbmap for the bucket and adds the vbmap to the vbucket_map_history only
if the map is balanced.

Seperately when a node is gracefully failed-over and rebalanced-in
it expects (the way our code is written) to find a map in
vbucket_map_history key and on not finding one we fail with the
error - {error, delta_recovery_not_possible}.

The notion of balance was stricter when we used the max-flow vbmap
algorithm - both the replica counts and active counts on all nodes
can't differ by more than 1.

With the greedy vbmap algorithm the acceptable max difference in
replica count between any two nodes (replica count range) is 5.

This patch also introduces a new config knob (max_replica_count_range)
which can change the default "replica count range" from 5 to any
desired value, if necessary.

Change-Id: Iea0be0e603453304fe0849eae22403224cdde47c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/176557
Well-Formed: Restriction Checker
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>