MB-54134: [BP] fix secrets leak in REST api crash
Fixes a leak of basic auth user/password tuple when we encounter an
error in certain REST api requests. The ones identified for this
related to the addition of a server to a group. There may be other
places that are leaked in a similar fashion but I was unable to find
any specific places.

Change-Id: Ia9a3fb836fe3c4dcdb900a4e534d55e4a419e5f3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/183091
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>