MB-55427: Fix set manifest modify_collection bug and refactor test
This test tests a very narrow portion of the set manifest test path,
namely it tests that we calculate the correct operations to perform
based on the existing manifest and the desired manifest. It does not
take into account the results of any individual operations though and is
too narrow in scope to add meaningful value. Refactor the test to
perform the bulk of the set manifest path to test that the resulting
manifest after operations are performed is correct.

This test refactor hits a bug in the modify_collection path which was
previously not using the collection props. This would immediately hit a
function_clause error as the verify_oper overload for modify_collection
expects that the collection props exist.

It may be noted that some of the history values are incorrectly set
here. When history is specified as false the value is ignored by the set
manifest path as we remove defaults when we attempt to get the
operations to perform. This is a bug which will be addressed in a
followup as the primary purpose of this change is to refactor the test;
changing the code to fix the modify_collection path was just
unfortunately unavoidable.

Change-Id: Ibb2a015ce41a0c994806ffaf457d84adff04c8df
Reviewed-on: https://review.couchbase.org/c/ns_server/+/186307
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>